big changes)))

+++ SushiShop.Core/Data/Dtos/Toppings/ToppingDto.cs
@@ -4,10 +4,10 @@ namespace SushiShop.Core.Data.Dtos.Toppings
 {
     public class ToppingDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [JsonProperty("pagetitle")]
         public string? PageTitle { get; set; }
-        public long Price { get; set; }
+        public decimal Price { get; set; }
         [JsonProperty("toppingCategory")]
         public string? ToppingCategory { get; set; }
         public int CountInBasket { get; set; }
