+++ SushiShop.Core/Services/Http/Products/ProductsService.cs
@@ -18,7 +18,7 @@ namespace SushiShop.Core.Services.Http.Products
         }
 
         public Task<HttpResponse<ResponseDto<ProductDto>>> GetProductAsync(
-            int id,
+            long id,
             string? city,
             Guid? basketId,
             CancellationToken cancellationToken)
@@ -38,7 +38,7 @@ namespace SushiShop.Core.Services.Http.Products
         }
 
         public Task<HttpResponse<ResponseDto<ProductDto[]>>> GetProductsByCategoryAsync(
-            int? categoryId,
+            long? categoryId,
             string? city,
             Guid? basketId,
             StickerType? stickerType,
@@ -60,7 +60,7 @@ namespace SushiShop.Core.Services.Http.Products
         }
 
         public Task<HttpResponse<ResponseDto<ProductDto[]>>> GetRelatedProductsAsync(
-            int id,
+            long id,
             string? city,
             Guid? basketId,
             CancellationToken cancellationToken)
